Move JSON uploader controller to its own file
